Add an (optional) mapping between field names
This enables us to write:

    type t = { foo: int } with rpc("foo" => "type")

This means that in ML, you will write x.foo, but this will be translated to a JSON record where the field is named "type".

No sanity checks are done at all, so it the developper's responsability to keep the mapping bijective and to not make typo!