Fix struct_extend handling of option types
If any field has default value None, it will not be present at all in
the rpc. This field will therefore get skipped when mapping over the
default fields, and the extended record will always have this field set
to None.

Instead, fold across the default fields, and add any which are missing
to the rpc which we are extending.

Signed-off-by: John Else <john.else@citrix.com>