Add defaults for polymorphic variants.
If the Rpc.t we're unmarshalling looks like it could be a valid
polymorphic variant (ie, it looks like a Rpc.String or an
Rpc.Enum [Rpc.String; ...]), then we return the default value.

This additionally fixes an issue when trying to marshal something
of type (<polyvar> * t) list. If the polyvar is representable as
an Rpc.String (ie, it has no extra values associated), then we
will now represent that as a Dict rather than as an Enum.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>