Fix marshalling of optional arguments
Previously if an argument to an RPC call was a named optional value it
would be unconditionally added to the dictionary of params and the
optionness encoded by either passing an empty or one element Enum in
the dict. This patch makes it more consistent with how optional
record values are encoded - if 'None' the names argument is now
omitted from the arg dict, and if 'Some' the value is unboxed from the
Enum and put directly into the dict.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>