Avoid stack overflow due to List.map in Jsonrpc, Rpcmarshal, ppx
Use List.rev_map and List.rev instead of List.map to avoid a stack
overflow when unmarshalling a large list with many elements. I call
List.rev in case of Dicts as well, to avoid breaking users that rely on
the original ordering, even though preserving the order is not strictly
necessary in this case. I've extracted these into a
Rpcmarshal.tailrec_map function that is called from the generated ppx
code. I didn't convert the List.maps in case of variants and structs,
because these cannot be so big that they cause a stack overflow.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>