Make sure fns are added with correct name in GenServer
The standard use case for the IDL is that we declare the RPC functions
in a functor taking an RPC module, and set the description in  that
module. So at that point when we call declare, the description may not
be set. When a user of the interface supplies the implementations of the
RPC functions, the description must be available, and therefore we know
the final wire name. Previously we constructed the wire name when the
description was still unset, resulting in an incorrect wire name or an
exception.

In the Lwt and Async servers, the functions were not added with their
wire names that include the namespace, and they were also missing the
check for missing interface description, so I've fixed that.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>