markdowngen: document nested variants and structs
These variants and structs that are not an input/output parameter, and
therefore not included in type_decls, also contain useful documentation
for their tags / fields, so it's helpful to document these too.
Now these variants and structs will be referenced by name in type
definitions, since their documentation is included in the page.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>