rpcty ppx: Avoid warning 23: "with clause is useless"
When setting the record field of a record that has only one field, avoid
this warning, by omitting the "with" clause:

Warning 23: all the fields are explicitly listed in this record:
the 'with' clause is useless.

I've updated the flags of the PPX test suite so that now it would fail
if any warning appeared. This change is useful for projects that have
enabled all warnings as errors.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>