markdowngen: document params with same types but different defs
Sometimes we create an alias to a simple type, like a string, and add
a custom name and description, for the sake of documenting a function
parameter. Previously these were filtered out, because their types were
already documented. Now this won't happen because definitions are only
compared based on their names. I've made sure that we do not duplicate
the already documented function parameters when we expand their nested
types.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>