pythongen: fix generated skeleton class to pass typechecking
The behaviour of the X_test Python classes did not match the generated
type checking code. If the return parameter is named, the test class
will put the return value into a dictionary under that name, however,
the typechecking code will still expect the return value itself, not a
dictionary. I think the behaviour of the type checking code is the
correct one, because there is no need to name the returned value,
because we only return one value.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>