pythongen: generate classes for errors in interface
If an error variant in the interface has the same name as one of the
errors used by the generated code, then we won't generate code for that
error, to avoid having two Python classes with the same name. This
allows one to catch the errors defined by the helpers, provided they
have a compatible definition in the interface.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>