pythongen test: make pylint checks stricter
Instead of only checking for errors, now we have a whitelist of
warnings.

I had to rename TypeError to RpcTypeError to ensure it does not class
with the built-in exception with the same name. It does not seem to be
used by clients, SM raises the normal TypeError at a few places, but
that should be fine.

Warnings I did not remove:
* line-too-long,too-few-public-methods(for the exns),unused-argument(for
  the skeleton classes),no-self-use(for the skeletons, ...):
  it's pointless or impossible to remove these
* invalid-name: We cannot convert the classes to CamelCase without breaking the interface
* broad-except: it is necessary to catch all exceptions at some places
* protected-access: I did not want to make _dispatcher public for now
* redefined-builtin(due to TypeError): again, this would be a breaking
  change because of the helper module that can be passed in - I'm planning
  to remove that later

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>