add option in jsonrpc.of_string to ignore tailing junk from input
There is need to just retrieve the valid json from the input string,
and convert the json to RPC object. Tailing junks need to be ignored
to enhance robustness of the application. Without the changes, user
need to parse the input using some json parser, and then make the
string (according to the parser) to be parsed again here in Yojson
inside jsonrpc, which makes the API inefficient to user.

Signed-off-by: YarsinCitrix <yarsin.he@citrix.com>