Idl.declare: check for duplicate argument names
We've recently had a bug where an RPC got declared with duplicate names
due to a confusion on how [Param.mk] works: the string argument taken by
[Param.mk] is NOT a type name, but an argument name.
Which then resulted in the RPC silently swapping the order of the arguments around.

Prevent this mistake at startup time and reject any RPC with duplicate
parameter names declared, it is a programming error.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>