Fix unmarshaling datetime and base64 for JSONRPC
JSONRPC marshals these as string, but then trying to unmarshal would raise an exception.
We need to accept String when unmarshaling DateTime and Base64 too.

Some type safety is lost, because JSON doesn't have a way to distinguish between these (XMLRPC does).

Update the unit test to check that this now works.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>