xmlrpc: optimize buffered encoding
Encode directly to a buffer where possible, and avoid constructing intermediate strings.
This isn't always beneficial (benchmarks on 4.14.2 show that Int64.to_string is slower than bprintf,
but Int32.to_string is faster).
When encoding parameters it is useful to avoid encoding to intermediate strings though.

Optimize Internals.encode, by specializing it to the 'translate' function used here.
The default OCaml inliner isn't able to inline 'translate', even when '[@inline always]' is used,
and even when both functions are moved to the same file.

Write a specialized version for xmlrpc that doesn't take translate as a parameter.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>